+++ src/main/java/com/napier/courswork/City.java
@@ -5,7 +5,7 @@ public class City {
     public String Name;
     public String CountryCode;
     public String District;
-    public int Population;
+    public Long Population;
     public String Region;
     public String Continent;
     public String Capital;
